# Add DecimalSelectiveReader