```
===NEW===
Benchmark                                       (typeSignature)  (withNulls)  Mode  Cnt  Score    Error  Units
BenchmarkSelectiveStreamReaders.read              decimal(10,5)         true  avgt   20  0.234 ±  0.008   s/op
BenchmarkSelectiveStreamReaders.read              decimal(10,5)        false  avgt   20  0.471 ±  0.015   s/op
BenchmarkSelectiveStreamReaders.readAllNull       decimal(10,5)          N/A  avgt   20  0.002 ±  0.001   s/op
BenchmarkSelectiveStreamReaders.readWithFilter    decimal(10,5)         true  avgt   20  0.277 ±  0.009   s/op
BenchmarkSelectiveStreamReaders.readWithFilter    decimal(10,5)        false  avgt   20  0.466 ±  0.018   s/op

Benchmark                                             Mode  Cnt  Score   Error  Units
BenchmarkBatchStreamReaders.readShortDecimalNoNull    avgt   60  0.481 ± 0.009   s/op
BenchmarkBatchStreamReaders.readShortDecimalWithNull  avgt   60  0.301 ± 0.005   s/op

Benchmark                                       (typeSignature)  (withNulls)  Mode  Cnt  Score    Error  Units
BenchmarkSelectiveStreamReaders.read             decimal(30,10)         true  avgt   20  0.288 ±  0.008   s/op
BenchmarkSelectiveStreamReaders.read             decimal(30,10)        false  avgt   20  0.446 ±  0.025   s/op
BenchmarkSelectiveStreamReaders.readAllNull      decimal(30,10)          N/A  avgt   20  0.002 ±  0.001   s/op
BenchmarkSelectiveStreamReaders.readWithFilter   decimal(30,10)         true  avgt   20  0.350 ±  0.011   s/op
BenchmarkSelectiveStreamReaders.readWithFilter   decimal(30,10)        false  avgt   20  0.585 ±  0.036   s/op

Benchmark                                            Mode  Cnt  Score   Error  Units
BenchmarkBatchStreamReaders.readLongDecimalNoNull    avgt   60  0.645 ± 0.013   s/op
BenchmarkBatchStreamReaders.readLongDecimalWithNull  avgt   60  0.483 ± 0.010   s/op

===OLD===
Benchmark                                       (typeSignature)  (withNulls)  Mode  Cnt  Score    Error  Units
BenchmarkSelectiveStreamReaders.read              decimal(10,5)         true  avgt   20  0.255 ±  0.008   s/op
BenchmarkSelectiveStreamReaders.read              decimal(10,5)        false  avgt   20  0.465 ±  0.022   s/op
BenchmarkSelectiveStreamReaders.readAllNull       decimal(10,5)          N/A  avgt   20  0.002 ±  0.001   s/op
BenchmarkSelectiveStreamReaders.readWithFilter    decimal(10,5)         true  avgt   20  0.466 ±  0.013   s/op


# Run complete. Total time: 00:04:09

Benchmark                                             Mode  Cnt  Score   Error  Units
BenchmarkBatchStreamReaders.readShortDecimalNoNull    avgt   60  0.490 ± 0.014   s/op
BenchmarkBatchStreamReaders.readShortDecimalWithNull  avgt   60  0.344 ± 0.026   s/op



BenchmarkSelectiveStreamReaders.read             decimal(30,10)         true  avgt   20  0.385 ±  0.009   s/op
BenchmarkSelectiveStreamReaders.read             decimal(30,10)        false  avgt   20  0.611 ±  0.012   s/op
BenchmarkSelectiveStreamReaders.readAllNull      decimal(30,10)          N/A  avgt   20  0.003 ±  0.001   s/op
BenchmarkSelectiveStreamReaders.readWithFilter   decimal(30,10)         true  avgt   20  0.430 ±  0.011   s/op
BenchmarkSelectiveStreamReaders.readWithFilter   decimal(30,10)        false  avgt   20  0.678 ±  0.008   s/op


Benchmark                                            Mode  Cnt  Score   Error  Units
BenchmarkBatchStreamReaders.readLongDecimalNoNull    avgt   60  0.636 ± 0.009   s/op
BenchmarkBatchStreamReaders.readLongDecimalWithNull  avgt   60  0.437 ± 0.013   s/op
```
====NO RELEASE NOTES====